Refactored some code for Auth0 Implementation. End result was creating a helpers folder where functions that aren't middleware, models or routes could live off of the Okta file as a dependency.
