+++ api/applications/applicationRouter.js
@@ -12,7 +12,7 @@ const {
 } = require('../middleware/applicationMiddleware');
 const { createProfile } = require('../middleware/profilesMiddleware');
 const { registerOktaUser } = require('../middleware/oktaAuth');
-const validation = require('../middleware/validation');
+const validation = require('../helpers/validation');
 const applicationSchema = require('../validations/application/applicationSchema');
 const config = require('../../config/dsConfig');
 const axios = require('axios');
