+++ api/helpers/passGenerator.js
@@ -1,0 +1,12 @@
+function passGenerator() {
+  let password = '';
+  const str =
+    'ABCDEFGHIJKLMNOPQRSTUVWXYZ' + 'abcdefghijklmnopqrstuvwxyz0123456789@#$';
+  for (let i = 1; i <= 8; i++) {
+    const char = Math.floor(Math.random() * str.length + 1);
+    password += str.charAt(char);
+  }
+  return password;
+}
+
+module.exports = { passGenerator };
