+++ api/middleware/oktaAuth.js
@@ -1,21 +1,11 @@
 const okta = require('@okta/okta-sdk-nodejs');
 const { config } = require('../../config/okta');
+const { passGenerator } = require('../helpers/passGenerator');
 
 const client = new okta.Client({
   orgUrl: config.orgUrl,
   token: config.devToken,
 });
-
-function passGenerator() {
-  let password = '';
-  const str =
-    'ABCDEFGHIJKLMNOPQRSTUVWXYZ' + 'abcdefghijklmnopqrstuvwxyz0123456789@#$';
-  for (let i = 1; i <= 8; i++) {
-    const char = Math.floor(Math.random() * str.length + 1);
-    password += str.charAt(char);
-  }
-  return password;
-}
 
 const registerOktaUser = async (req, res, next) => {
   const tempPassword = passGenerator();
